backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "alexandros",email = "alex.kyr.triant@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pdf2image (>=1.17.0,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "langchain-community (>=0.3.31,<0.4.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "langchain-ollama (>=0.3.10,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "chromadb (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
